Move our types into patches.rs
